+++ src/Testcontainers/Clients/DockerNetworkOperations.cs
@@ -62,7 +62,7 @@ namespace DotNet.Testcontainers.Clients
         Name = configuration.Name,
         Driver = configuration.Driver.Value,
         Labels = configuration.Labels.ToDictionary(item => item.Key, item => item.Value),
-        Options = configuration.Options?.ToDictionary(item => item.Key, item => item.Value),
+        Options = configuration.Options.ToDictionary(item => item.Key, item => item.Value),
       };
 
       var id = (await this.Docker.Networks.CreateNetworkAsync(createParameters, ct)