+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - 640 Add `ITestcontainersBuilder<TDockerContainer>.WithResourceMapping` to copy files or or any binary contents into the created container even before it is started.
+- 654 Add `ITestcontainersNetworkBuilder.WithOption` (@vlaskal)
 
 ### Changed